\name{NoC-overview}
\alias{NoC-overview}
\alias{NoC}
\docType{package}
\title{
The Number of Classes (NoC) Package
}
\description{
The Number of Classes package contains a function that estimates the total number of classes of a kind, given the discovery rate of known classes and a numerical proxy of effort.
}
\details{
\tabular{ll}{
Package: \tab NoC\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2024-11-13\cr
License: \tab GPL (>=3)\cr
}

}

\author{
Daniel Dempsey <\email{daniel.dempsey0@gmail.com}>, with contributions and supervision by Simon Wilson <\email{swilson@tcd.ie}>
}

\references{
TBD
}

\keyword{package}

\seealso{
\code{\link{CD_ABC}}
}

